requirements.txt
beautifulsoup4==4.10.0
certifi==2024.2.2
charset-normalizer==2.0.12
idna==3.7
pillow==10.3.0
PyPDF2==3.0.1
requests==2.26.0
soupsieve==2.5
urllib3==1.26.18
